pyproject.toml
[tool.poetry]
name = "appointment-scheduling"
version = "0.1.0"
description = "Template Django project configured with Postgres"
authors = ["shikhardslr <shikhar.prasad@dslrteam.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
psycopg = {extras = ["binary"], version = "^3.1"}
dj-database-url = "^2.1"
python-dotenv = "^1.0"
djangorestframework = "^3.15"
django-cors-headers = "^4.4"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
pytest = "^8.3"
pytest-django = "^4.8"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
